+++ tests/diagnostic_base/test_base_util.py
@@ -247,9 +247,15 @@ def test_start_end_dates():
         pd.Timestamp("2020-01-02"),
     )
 
-    assert start_end_dates(start_std="2020-01-01", end_std="2020-01-02") == (None, None)
+    assert start_end_dates(start_std="2020-01-01", end_std="2020-01-02") == (
+        pd.Timestamp("2020-01-01"),
+        pd.Timestamp("2020-01-02"),
+    )
 
-    assert start_end_dates(startdate="2020-01-01", end_std="2020-01-02") == (pd.Timestamp("2020-01-01"), None)
+    assert start_end_dates(startdate="2020-01-01", end_std="2020-01-02") == (
+        pd.Timestamp("2020-01-01"),
+        pd.Timestamp("2020-01-02"),
+    )
 
 
 @pytest.mark.parametrize(